+++ lua/obsidian/util.lua
@@ -3,6 +3,42 @@ local echo = require "obsidian.echo"
 
 local util = {}
 
+---@enum OSType
+util.OSType = {
+  Linux = "Linux",
+  Wsl = "Wsl",
+  Windows = "Windows",
+  Darwin = "Darwin",
+}
+
+util._current_os = nil
+
+---Get the running operating system.
+---Reference https://vi.stackexchange.com/a/2577/33116
+---@return OSType
+util.get_os = function()
+  if util._current_os ~= nil then
+    return util._current_os
+  end
+
+  local this_os
+  if vim.fn.has "win32" == 1 then
+    this_os = util.OS.Windows
+  else
+    local sysname = vim.loop.os_uname().sysname ---@diagnostic disable-line: undefined-field
+    local release = vim.loop.os_uname().release ---@diagnostic disable-line: undefined-field
+    if sysname == "Linux" and string.find(release, "microsoft") then
+      this_os = util.OS.Wsl
+    else
+      this_os = sysname
+    end
+  end
+
+  assert(this_os)
+  util._current_os = this_os
+  return this_os
+end
+
 ---Get the strategy for opening notes
 ---
 ---@param opt "current"|"vsplit"|"hsplit"
@@ -939,6 +975,52 @@ util.get_named_buffers = function()
       return bufnr, vim.fs.normalize(vim.api.nvim_buf_get_name(bufnr))
     end
   end
+end
+
+---Insert text at current cursor position.
+---@param text string
+util.insert_text = function(text)
+  local curpos = vim.fn.getcurpos()
+  local line_num, line_col = curpos[2], curpos[3]
+  local indent = string.rep(" ", line_col)
+
+  -- Convert text to lines table so we can handle multi-line strings.
+  local lines = {}
+  for line in text:gmatch "[^\r\n]+" do
+    lines[#lines + 1] = line
+  end
+
+  for line_index, line in pairs(lines) do
+    local current_line_num = line_num + line_index - 1
+    local current_line = vim.fn.getline(current_line_num)
+    assert(type(current_line) == "string")
+
+    -- Since there's no column 0, remove extra space when current line is blank.
+    if current_line == "" then
+      indent = indent:sub(1, -2)
+    end
+
+    local pre_txt = current_line:sub(1, line_col)
+    local post_txt = current_line:sub(line_col + 1, -1)
+    local inserted_txt = pre_txt .. line .. post_txt
+
+    vim.fn.setline(current_line_num, inserted_txt)
+
+    -- Create new line so inserted_txt doesn't replace next lines
+    if line_index ~= #lines then
+      vim.fn.append(current_line_num, indent)
+    end
+  end
+end
+
+---@param bufnr integer
+---@return string
+util.buf_get_full_text = function(bufnr)
+  local text = table.concat(vim.api.nvim_buf_get_lines(bufnr, 0, -1, true), "\n")
+  if vim.api.nvim_get_option_value("eol", { buf = bufnr }) then
+    text = text .. "\n"
+  end
+  return text
 end
 
 return util